watchtower_v2/requirements.txt
flask==3.0.2
flask-cors==4.0.0
torch==2.2.0
torchvision==0.17.0
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
pytesseract==0.3.10 